fix: correct int field parsing mistakenly changed to long
Fixes false positives from bulk instanceId int->long migration:
- count (GameObjectTools.cs)
- maxDepth (GetGameObjectTool.cs, GetGameObjectResource.cs)
- newLayer (UpdateGameObjectTool.cs)
- slot (MaterialTools.cs)

These are regular int parameters, not EntityIds.

+++ Editor/Resources/GetGameObjectResource.cs
@@ -87,7 +87,7 @@ namespace McpUnity.Resources
                 };
             }
 
-            int maxDepth = parameters["maxDepth"]?.ToObject<long?>() ?? DefaultMaxChildDepth;
+            int maxDepth = parameters["maxDepth"]?.ToObject<int?>() ?? DefaultMaxChildDepth;
             bool includeComponents = parameters["includeComponents"]?.ToObject<bool?>() ?? true;
             bool includeComponentProperties = parameters["includeComponentProperties"]?.ToObject<bool?>() ?? true;
