+++ Editor/Tools/UpdateGameObjectTool.cs
@@ -35,7 +35,7 @@ namespace McpUnity.Tools
 
             string newName = gameObjectData? ["name"]?.ToObject<string>();
             string newTag = gameObjectData? ["tag"]?.ToObject<string>();
-            int? newLayer = gameObjectData? ["layer"]?.ToObject<long?>();
+            int? newLayer = gameObjectData? ["layer"]?.ToObject<int?>();
             bool? newIsActiveSelf = (gameObjectData?["activeSelf"] ?? gameObjectData?["isActiveSelf"])?.ToObject<bool?>();
             bool? newIsStatic = (gameObjectData?["isStatic"] ?? gameObjectData?["static"])?.ToObject<bool?>();
